+++ lib/src/outbox/outbox-listener.ts
@@ -37,5 +37,6 @@ export const initializeOutboxListener = (
     logErrors,
     logger,
     strategies,
+    'outbox',
   );
 };